fixed with the new decrypt and encrypt files

+++ decrypt.c
@@ -1,7 +1,7 @@
-void decrypt (long *v, long *k){
+void decrypt (int *v, int *k){
 /* TEA decryption routine */
-unsigned long n=32, sum, y=v[0], z=v[1];
-unsigned long delta=0x9e3779b9l;
+unsigned int n=32, sum, y=v[0], z=v[1];
+unsigned int delta=0x9e3779b9l;
 
 	sum = delta<<5;
 	while (n-- > 0){
